backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
defusedxml==0.8.0rc2
Django==3.2.3
django-colorfield==0.11.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.1.0
